requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.12.0
numpy>=1.21.0
scipy>=1.10.0
pandas>=1.5.0
Pillow>=9.0.0
tqdm>=4.65.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
jupyter>=1.0.0
pyyaml>=6.0
tensorboard>=2.12.0 